Move project and name record queries to queries dir
