+++ packages/truffle-db/src/loaders/queries/index.ts
@@ -4,3 +4,5 @@ export { AddCompilation } from "./compilation";
 export { AddContracts } from "./contracts";
 export { AddContractInstances } from "./contractInstances";
 export { AddNetworks } from "./networks";
+export { AddNameRecords } from "./nameRecords";
+export { GetProject, AddProject, SetProjectNames } from "./project";
